services/planning-grpc/server/requirements.txt
grpcio>=1.76.0
grpcio-tools>=1.76.0
